MAR-53 move docs to app dir
