--- src/app/[slug]/page.tsx
@@ -1,67 +1,0 @@
-import { getPostBySlug, getAllPosts } from '@/utils/api';
-type Author = {
-  name: string;
-  picture: string;
-};
-type PostType = {
-  slug: string;
-  title: string;
-  date: string;
-  coverImage: string;
-  author: Author;
-  excerpt: string;
-  ogImage: {
-    url: string;
-  };
-  content: string;
-};
-type Props = {
-  post: PostType;
-  morePosts: PostType[];
-  preview?: boolean;
-};
-
-export default function Post({ post }: Props) {
-  return <div>{JSON.stringify(post)}</div>;
-}
-
-type Params = {
-  params: {
-    slug: string;
-  };
-};
-
-export async function getStaticProps({ params }: Params) {
-  const post = getPostBySlug(params.slug, [
-    'title',
-    'date',
-    'slug',
-    'author',
-    'content',
-    'ogImage',
-    'coverImage',
-  ]);
-
-  return {
-    props: {
-      post: {
-        ...post,
-      },
-    },
-  };
-}
-
-export async function getStaticPaths() {
-  const posts = getAllPosts(['slug']);
-
-  return {
-    paths: posts.map((post) => {
-      return {
-        params: {
-          slug: post.slug,
-        },
-      };
-    }),
-    fallback: false,
-  };
-}
