+++ src/app/docs/page.tsx
@@ -1,0 +1,5 @@
+import { cwd } from 'process';
+
+export default function Docs() {
+  return cwd();
+}
