--- src/pages/_app.tsx
@@ -1,6 +1,0 @@
-import { AppProps } from 'next/app';
-import '../app/globals.css';
-
-export default function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
-}
