--- src/pages/components/mdComponent.tsx
@@ -1,12 +1,0 @@
-export function Heading(props: string) {
-  return <h1>{props}</h1>;
-}
-export function H2(props: string) {
-  return <h2>{props}</h2>;
-}
-export function H3(props: string) {
-  return <h3>{props}</h3>;
-}
-export function Text(props: string) {
-  return <p>{props}</p>;
-}
